+++ ibc/path/keys_ics02.go
@@ -1,4 +1,4 @@
-package host
+package path
 
 import "fmt"
