+++ ibc/path/keys_ics04.go
@@ -1,4 +1,4 @@
-package host
+package path
 
 import (
 	"fmt"
