+++ ibc/path/path_test.go
@@ -1,4 +1,4 @@
-package host
+package path
 
 import (
 	"testing"
