+++ ibc/types/convert.go
@@ -1,19 +1,18 @@
-package ibc
+package types
 
 import (
 	"fmt"
 
-	"github.com/pokt-network/pocket/ibc/types"
 	"github.com/pokt-network/pocket/shared/codec"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/crypto"
 	"google.golang.org/protobuf/types/known/anypb"
 )
 
-func CreateUpdateStoreMessage(key, value []byte) *types.IBCMessage {
-	return &types.IBCMessage{
-		Event: &types.IBCMessage_Update{
-			Update: &types.UpdateIBCStore{
+func CreateUpdateStoreMessage(key, value []byte) *IBCMessage {
+	return &IBCMessage{
+		Event: &IBCMessage_Update{
+			Update: &UpdateIBCStore{
 				Key:   key,
 				Value: value,
 			},
@@ -21,23 +20,23 @@ func CreateUpdateStoreMessage(key, value []byte) *types.IBCMessage {
 	}
 }
 
-func CreatePruneStoreMessage(key []byte) *types.IBCMessage {
-	return &types.IBCMessage{
-		Event: &types.IBCMessage_Prune{
-			Prune: &types.PruneIBCStore{
+func CreatePruneStoreMessage(key []byte) *IBCMessage {
+	return &IBCMessage{
+		Event: &IBCMessage_Prune{
+			Prune: &PruneIBCStore{
 				Key: key,
 			},
 		},
 	}
 }
 
-func ConvertIBCMessageToTx(ibcMessage *types.IBCMessage) (*coreTypes.Transaction, error) {
+func ConvertIBCMessageToTx(ibcMessage *IBCMessage) (*coreTypes.Transaction, error) {
 	var anyMsg *anypb.Any
 	var err error
 	switch event := ibcMessage.Event.(type) {
-	case *types.IBCMessage_Update:
+	case *IBCMessage_Update:
 		anyMsg, err = codec.GetCodec().ToAny(event.Update)
-	case *types.IBCMessage_Prune:
+	case *IBCMessage_Prune:
 		anyMsg, err = codec.GetCodec().ToAny(event.Prune)
 	default:
 		return nil, coreTypes.ErrIBCUnknownMessageType(fmt.Sprintf("%T", event))
